modularization: combine tv-shows data and ui
