+++ feature/tv-shows/build.gradle
@@ -32,9 +32,7 @@ android {
 }
 
 dependencies {
-    implementation project(':shared:components')
-    implementation project(':shared:data:tv')
-    implementation project(':shared:domain')
+    implementation project(':core')
 
     implementation libraries.coreKtx
     implementation libraries.appcompat
@@ -46,7 +44,11 @@ dependencies {
 
     debugImplementation libraries.leakCanary
 
-    implementation libraries.rxJava
-
     implementation libraries.timber
+
+    implementation libraries.retrofit
+
+    implementation libraries.gson
+
+    implementation libraries.rxJava
 }
