+++ feature/tv-shows/src/main/java/com/my/tv_shows/data/di/TvFactory.kt
@@ -1,10 +1,10 @@
-package com.my.tv.di
+package com.my.tv_shows.data.di
 
-import com.my.domain.repository.TvRepository
-import com.my.remote.RemoteFactory
-import com.my.tv.TvRepositoryImpl
-import com.my.tv.remote.TvApi
-import com.my.tv.remote.TvRemoteDataSource
+import com.my.core.data.RemoteFactory
+import com.my.tv_shows.data.TvRepositoryImpl
+import com.my.tv_shows.data.remote.TvApi
+import com.my.tv_shows.data.remote.TvRemoteDataSource
+import com.my.tv_shows.domain.TvRepository
 
 class TvFactory {
     fun provideRepository(): TvRepository {
