+++ feature/tv-shows/src/main/java/com/my/tv_shows/di/TvShowsFactory.kt
@@ -1,7 +1,7 @@
 package com.my.tv_shows.di
 
-import com.my.domain.usecase.FetchPopularTvShowsUseCase
-import com.my.tv.di.TvFactory
+import com.my.tv_shows.data.di.TvFactory
+import com.my.tv_shows.domain.FetchPopularTvShowsUseCase
 import com.my.tv_shows.presentation.TvShowsPresenter
 
 class TvShowsFactory {
