+++ feature/tv-shows/src/main/java/com/my/tv_shows/presentation/TvShowsView.kt
@@ -1,7 +1,7 @@
 package com.my.tv_shows.presentation
 
-import com.my.domain.entity.TvShowsEntity
-import com.my.resources.mvp.BaseView
+import com.my.core.mvp.BaseView
+import com.my.tv_shows.domain.TvShowsEntity
 
 interface TvShowsView : BaseView {
     fun setTvShowsList(content: List<TvShowsEntity>)
