--- shared/data/tv/build.gradle
@@ -1,44 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-}
-
-android {
-    compileSdkVersion versions.compileSdk
-    buildToolsVersion versions.buildTools
-
-    defaultConfig {
-        minSdkVersion versions.minSdk
-        targetSdkVersion versions.targetSdk
-        versionCode 1
-        versionName "1.0"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-}
-
-dependencies {
-    implementation project(':shared:data:remote')
-    implementation project(':shared:domain')
-
-    implementation libraries.coreKtx
-    implementation libraries.appcompat
-    implementation libraries.material
-
-    implementation libraries.retrofit
-
-    implementation libraries.gson
-
-    implementation libraries.rxJava
-}
